Task 4 context_aware_rag_chatbot/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-qdrant
langchain-groq

qdrant-client

sentence-transformers

streamlit

python-dotenv

tiktoken


